[refactor] Move `ModelListFeatureTests` to single file
