+++ Features/Tests/FeaturesTests/FeaturesTests.swift
@@ -52,36 +52,6 @@ extension XCTestCase {
 }
 
 @MainActor
-final class ModelListFeatureTests: XCTestCase {
-  func test_settingsSheet_presentedWhenSettingButtonTapped() async {
-    let store = TestStore(
-      initialState: ModeListFeature.State(featureCards: FeatureCard.default)
-    ) {
-      ModeListFeature()
-    }
-
-    await store.send(.settingsButtonTapped) {
-      $0.presentSettingsPage = SettingsFeature.State()
-    }
-  }
-
-  func test_settingsSheet_dismissedWhenDoneButtonTapped() async {
-    let store = TestStore(
-      initialState: ModeListFeature.State(
-        presentSettingsPage: SettingsFeature.State(),
-        featureCards: FeatureCard.default
-      )
-    ) {
-      ModeListFeature()
-    }
-
-    await store.send(.settingsSheetDoneButtonTapped) {
-      $0.presentSettingsPage = nil
-    }
-  }
-}
-
-@MainActor
 final class UserSettingsFeatureTests: XCTestCase {
   func test_openURL_navigateToFeedbackForm() async {
     let store = TestStore(
